+++ packages/localizations/src/zh-CN.ts
@@ -602,6 +602,12 @@ export const zhCN: LocalizationResource = {
       subtitle: '继续使用 {{applicationName}}',
       title: '查看您的电子邮件',
     },
+    emailCodeMfa: {
+      formTitle: '查看您的电子邮件',
+      resendButton: '未收到验证码？重新发送',
+      subtitle: '继续使用 {{applicationName}}',
+      title: '查看您的电子邮件',
+    },
     emailLink: {
       clientMismatch: {
         subtitle: undefined,
@@ -637,6 +643,12 @@ export const zhCN: LocalizationResource = {
         titleNewTab: '在其他标签页上登录',
       },
     },
+    emailLinkMfa: {
+      formSubtitle: '使用发送到您电子邮件的验证链接',
+      resendButton: '没有收到链接？重新发送',
+      subtitle: '继续使用 {{applicationName}}',
+      title: '查看您的电子邮件',
+    },
     enterpriseConnections: {
       subtitle: undefined,
       title: undefined,
@@ -654,6 +666,7 @@ export const zhCN: LocalizationResource = {
       label__alternativeMethods: '或者，使用其他方式登录。',
       title: '忘记密码？',
     },
+    newDeviceVerificationNotice: '您正在从新设备登录。我们需要验证以保护您的账户安全。',
     noAvailableMethods: {
       message: '无法继续登录。没有可用的身份验证因素。',
       subtitle: '出现错误',